backend/userlist/serializers.py
from .models import List, User
from rest_framework import serializers

class ListSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = List
        fields = ['cliente', 'originador', 'url']


class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ['cliente']
